move buttonState and it's subclasses to a separate file
